Issue #429: WordPress plugin cleanup.
