+++ extras/wordpress/thinkup/help/Getting started.txt
@@ -28,7 +28,7 @@ The shortcodes currently available to you in this plugin are as follows:
 
 [thinkup_chronological_archive]
 
-[thinkup_status_replies post_id="post_id"]
+[thinkup_post_replies post_id="post_id"]
 
 [thinkup_reply_count post_id="post_id"]
 
@@ -36,7 +36,7 @@ The first shortcode does not require any arguments. The other two, however, requ
 
 Easiest first, [thinkup_reply_count post_id="1234"] will show you how many replies the post with a post_id of 1234 has. By default this will look for a Twitter post but you can change that by specifying the network attribute like so: [thinkup_reply_count post_id="1234" network="facebook"].
 
-You can display the replies for any post by using the [thinkup_status_replies post_id="1234"] shortcode. Again, this will search for Twitter psots by default but by specifying the network attribute you can change it to Facebook. 
+You can display the replies for any post by using the [thinkup_post_replies post_id="1234"] shortcode. Again, this will search for Twitter psots by default but by specifying the network attribute you can change it to Facebook. 
 
 And finally, you can view your ThinkUp's entire archive for any account by using the [thinkup_chronological_archive] shortcode. By default this will use the Twitter username specified in the settings and display Tweets. However, you can change this to your Facebook account by setting the username attribute to your full name on Facebook and the network attribute to "facebook", e.g. [thinkup_chronoglogical_archive network="facebook" username="Sam Rose"] will display my entire chronological archive of Facebook statuses.
